refactor(core): consolidate per-source test files (refs #213)
